GB-56 refactor : 로그인,로그아웃,토큰 발급 공통 로직
