+++ src/main/java/garlicbears/quiz/domain/management/user/controller/UserController.java
@@ -20,11 +20,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import garlicbears.quiz.domain.common.dto.ResponseDto;
-import garlicbears.quiz.domain.common.entity.Role;
 import garlicbears.quiz.domain.common.entity.User;
 import garlicbears.quiz.domain.common.service.LogService;
 import garlicbears.quiz.domain.management.common.dto.LoginDto;
 import garlicbears.quiz.domain.management.common.dto.ResponseUserDto;
+import garlicbears.quiz.domain.management.common.service.AuthService;
 import garlicbears.quiz.domain.management.user.dto.SignUpDto;
 import garlicbears.quiz.domain.management.user.dto.UpdateUserDto;
 import garlicbears.quiz.domain.management.user.service.UserRatingService;
@@ -52,6 +52,7 @@ public class UserController implements SwaggerUserController {
 	private final PasswordEncoder passwordEncoder;
 	private final RefreshTokenService refreshTokenService;
 	private final LogService logService;
+	private final AuthService authService;
 
 	@Autowired
 	UserController(UserService userService,
@@ -59,13 +60,15 @@ public class UserController implements SwaggerUserController {
 		JwtTokenizer jwtTokenizer,
 		PasswordEncoder passwordEncoder,
 		RefreshTokenService refreshTokenService,
-		LogService logService) {
+		LogService logService,
+		AuthService authService) {
 		this.userService = userService;
 		this.userRatingService = userRatingService;
 		this.jwtTokenizer = jwtTokenizer;
 		this.passwordEncoder = passwordEncoder;
 		this.refreshTokenService = refreshTokenService;
 		this.logService = logService;
+		this.authService = authService;
 	}
 
 	/**
@@ -190,36 +193,7 @@ public class UserController implements SwaggerUserController {
 			throw new CustomException(ErrorCode.INVALID_INPUT);
 		}
 
-		// 사용자의 역할(Role)을 문자열 리스트로 변환
-		List<String> roles = user.getRoles().stream()
-			.map(Role::getRoleName)
-			.toList();
-
-		// JWT토큰을 생성
-		String accessToken = jwtTokenizer.createAccessToken(user.getUserEmail(), user.getUserId(), roles);
-		String refreshToken = jwtTokenizer.createRefreshToken(user.getUserEmail(), user.getUserId(), roles);
-
-		// 리프레시 토큰을 Redis에 저장
-		refreshTokenService.save(user.getUserEmail(), refreshToken, JwtTokenizer.REFRESH_TOKEN_EXPIRE_COUNT);
-
-		// 리프레시 토큰을 쿠키에 저장
-		response.addCookie(createRefreshTokenCookie(refreshToken));
-
-		HttpHeaders headers = new HttpHeaders();
-		headers.set("Authorization", accessToken);
-
-		return ResponseEntity.ok().headers(headers).body(ResponseDto.success());
-	}
-
-	/**
-	 * 리프레시 토큰을 쿠키에 저장
-	 */
-	private Cookie createRefreshTokenCookie(String refreshToken) {
-		Cookie refreshTokenCookie = new Cookie("refreshToken", refreshToken);
-		refreshTokenCookie.setHttpOnly(true); // JavaScript에서 접근 불가능하도록 설정
-		refreshTokenCookie.setPath("/");
-		refreshTokenCookie.setMaxAge(7 * 24 * 60 * 60); // 7 days
-		return refreshTokenCookie;
+		return authService.createTokensAndRespond(user, response);
 	}
 
 	/**
@@ -227,30 +201,7 @@ public class UserController implements SwaggerUserController {
 	 */
 	@GetMapping("/reissue")
 	public ResponseEntity<?> requestRefresh(HttpServletRequest request, HttpServletResponse response) {
-		// 쿠키에서 리프레시 토큰을 읽기
-		String refreshToken = getRefreshTokenFromCookies(request);
-
-		// 전달받은 유저의 아이디로 유저가 존재하는지 확인
-		Claims claims = jwtTokenizer.parseRefreshToken(refreshToken);
-		String email = claims.getSubject();
-
-		// 전달받은 이메일로 리프레시 토큰이 존재하는지 확인하고, 존재하지 않으면 예외를 발생
-		refreshTokenService.findRefreshToken(email)
-			.orElseThrow(() -> new CustomException(ErrorCode.REFRESH_TOKEN_NOT_FOUND));
-
-		long userId = Long.valueOf((Integer)claims.get("id"));
-		List roles = (List)claims.get("roles");
-
-		// Token을 발급
-		String accessToken = jwtTokenizer.createAccessToken(email, userId, roles);
-		String newRefreshToken = jwtTokenizer.createRefreshToken(email, userId, roles);
-
-		// 기존 리프레시 토큰 삭제 후 새로운 리프레시 토큰 저장
-		refreshTokenService.deleteRefreshToken(email);
-		refreshTokenService.save(email, refreshToken, JwtTokenizer.REFRESH_TOKEN_EXPIRE_COUNT);
-
-		// 새로운 리프레시 토큰을 쿠키에 저장
-		response.addCookie(createRefreshTokenCookie(newRefreshToken));
+		String accessToken = authService.requestRefresh(request, response);
 
 		HttpHeaders headers = new HttpHeaders();
 		headers.set("Authorization", accessToken);
@@ -258,31 +209,6 @@ public class UserController implements SwaggerUserController {
 		return ResponseEntity.ok().headers(headers).body(ResponseDto.success());
 	}
 
-	/**
-	 * 쿠키에서 리프레시 토큰을 읽음
-	 */
-	private String getRefreshTokenFromCookies(HttpServletRequest request) {
-		Cookie[] cookies = request.getCookies();
-		if (cookies != null) {
-			for (Cookie cookie : cookies) {
-				if ("refreshToken".equals(cookie.getName())) {
-					return cookie.getValue();
-				}
-			}
-		}
-		throw new CustomException(ErrorCode.REFRESH_TOKEN_NOT_FOUND);
-	}
-
-	/**
-	 * 리프레시 토큰을 쿠키에서 삭제하는 메서드.
-	 */
-	private Cookie deleteRefreshTokenCookie() {
-		Cookie refreshTokenCookie = new Cookie("refreshToken", null);
-		refreshTokenCookie.setHttpOnly(true); // JavaScript에서 접근 불가능하도록 설정
-		refreshTokenCookie.setPath("/");
-		refreshTokenCookie.setMaxAge(0); // 쿠키 삭제
-		return refreshTokenCookie;
-	}
 
 	/**
 	 * 로그아웃
@@ -290,14 +216,7 @@ public class UserController implements SwaggerUserController {
 	 */
 	@DeleteMapping("/logout")
 	public ResponseEntity<?> logout(HttpServletRequest request, HttpServletResponse response) {
-		String refreshToken = getRefreshTokenFromCookies(request);
-		Claims claims = jwtTokenizer.parseRefreshToken(refreshToken);
-		String email = claims.getSubject();
-		// 리프레시 토큰 삭제
-		refreshTokenService.deleteRefreshToken(email);
-		// 쿠키에서 리프레시 토큰 삭제
-		response.addCookie(deleteRefreshTokenCookie());
-
+		authService.logout(request, response);
 		return ResponseEntity.ok(ResponseDto.success());
 	}
